Move getRepoKey util to shared utils
